% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_survival.R
\name{addAssignment}
\alias{addAssignment}
\title{Append assignment "a=b"}
\usage{
addAssignment(main = "", a, b)
}
\arguments{
\item{main}{current assignment string}

\item{a}{key of the assignment}

\item{b}{value of the assignment}
}
\value{
\code{main} with \code{a = b} appended with a comma as separator if
\code{main} is not empty
}
\description{
Append assignment "a=b"
}
