% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{catHeader}
\alias{catHeader}
\title{Print an underlined Header}
\usage{
catHeader(x, char)
}
\arguments{
\item{x}{caption}

\item{char}{character used for the underline}
}
\description{
Print an underlined Header
}
