% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{catStructure}
\alias{catStructure}
\title{Print the Structure of an Object}
\usage{
catStructure(x, max.level = NA)
}
\arguments{
\item{x}{any R object}

\item{max.level}{passed to \code{\link[utils]{capture.output}}}
}
\description{
Print the Structure of an Object
}
