% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{cbindDataFrames}
\alias{cbindDataFrames}
\title{Bind Columns from a list of Data Frames}
\usage{
cbindDataFrames(dataFrames, keyindex = 1:2)
}
\arguments{
\item{dataFrames}{list of data frames}

\item{keyindex}{indices of (key) columns to be excluded}
}
\description{
Bind Columns from a list of Data Frames
}
