% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompcal.R
\name{checkConvergence}
\alias{checkConvergence}
\title{Warn if Model did not converge for all Strata}
\usage{
checkConvergence(calibration, do.warn = TRUE)
}
\arguments{
\item{calibration}{calibration object as returned by
\code{kwbGompitz:::readCalibration}}

\item{do.warn}{if \code{TRUE} warnings are shown for non-calibrated strata}
}
\value{
named logical vector with element names corresponding to the names of
  the elements of \code{calibration} that represent the different strata.
}
\description{
Warn if the model did not converge for all strata
}
