% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputData.R
\name{checkInputData}
\alias{checkInputData}
\title{Check Input Data}
\usage{
checkInputData(input.data)
}
\arguments{
\item{input.data}{input.data}
}
\value{
error in case input data was not defined properly
}
\description{
Checks if input data is defined properly
}
