% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{columnwise}
\alias{columnwise}
\title{Apply a Function for each Column}
\usage{
columnwise(x, FUN, ...)
}
\arguments{
\item{x}{two dimensional object}

\item{FUN}{function to be called for each column}

\item{\dots}{arguments passed to \code{FUN}}
}
\description{
Apply a Function for each Column
}
