% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompcal.R
\name{compareEstimates}
\alias{compareEstimates}
\title{Compare Estimates in calib.txt and param.txt}
\usage{
compareEstimates(calibration, parameters, digits, warn = FALSE)
}
\arguments{
\item{calibration}{list structure as returned by
\code{kwbGompitz:::readCalibration}}

\item{parameters}{list structure as returned by
\code{kwbGompitz:::readParameters}}

\item{digits}{round the estimates to this number of significant (not 
decimal!) digits before comparing}

\item{warn}{if \code{TRUE} (the default is \code{FALSE}) a warning is given
if the strata read from \code{calibr.txt} do not correspond to the strata
read from \code{param.txt} (containing only the successfully calibrated
strata!?).}
}
\value{
number of warnings that occurred
}
\description{
Compare Estimates in calib.txt and param.txt
}
