% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{composeMasterData}
\alias{composeMasterData}
\title{Prepare "Master Data" for Gompitz Functions}
\usage{
composeMasterData(stratum, pipeid, instyear, inspyear, condition)
}
\arguments{
\item{stratum}{Stratum alphanumerical label}

\item{pipeid}{Pipeline Identifier alphanumerical label}

\item{instyear}{Installation year (integer)}

\item{inspyear}{Inspection Year (integer - void field if not inspected)}

\item{condition}{Condition Class alphanumerical label}
}
\value{
data.frame with columns \emph{stratum}, \emph{pipeid},
  \emph{instyear}, \emph{inspyear}, \emph{condition},
}
\description{
Prepare "master data" (stratum, pipe-ID, year of installation, year of 
inspection, condition class) for Gompitz functions
}
\seealso{
\code{\link{composeGompitzInputData}}
}
