% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{copyParameters}
\alias{copyParameters}
\title{Copy Model Parameters from one Stratum to another}
\usage{
copyParameters(calibration, from = NULL, to = NULL, dbg = TRUE)
}
\arguments{
\item{calibration}{calibration object as returned by
\code{\link{runGompitzCalibration}}}

\item{from}{name of stratum to copy parameters from}

\item{to}{vector of names of strata to which parameters are to be
copied to. If not given, the parameters of \code{from} are copied
to all strata for which model parameters did not converge.}

\item{dbg}{if \code{TRUE} (default) debug messages are shown}
}
\description{
Copy Model Parameters from one Stratum to another
}
\examples{
\dontrun{
# Get an example calibration
calibration <- kwbGompitz::exampleCalibration()

# Check for which strata the model parameters converged
checkConvergence(calibration, do.warn = FALSE)

# Copy parameters from a stratum for which the model parameters converged
# to a stratum for which the model parameters did not converge
calibration <- copyParameters(calibration, from = "Cast Iron", to = "Concrete")

# Check again
checkConvergence(calibration, do.warn = FALSE)

# The following gives a warning (no convergence for source stratum) and returns
# the calibration unchanged
calibration <- copyParameters(calibration, "Clay", "Brick")

# The following gives a warning (differing distinct condition classes) and
# returns the calibration unchanged
calibration <- copyParameters(calibration, "Cast Iron", "Brick")
}
}
