% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleCreate.R
\name{createExampleFiles}
\alias{createExampleFiles}
\title{Create Example Files}
\usage{
createExampleFiles()
}
\description{
Create Example Files Using Different Windows Executables. Run different
executables of gompcal and  gompred (as provided in subfolders 
"bin_<version>" with version in ["win32", "win32_kwb", "unix_home"]) on the 
example input file (obs.txt) and store the  results in folders 
"example_<version>".
}
