% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleCreate.R
\name{createExampleFilesSmall}
\alias{createExampleFilesSmall}
\title{Create smaller versions of the example files}
\usage{
createExampleFilesSmall(
  parts = c(2, 4, 8, 16),
  version = c("unix", "unix_home", "win32", "win32_kwb")[ifelse(kwb.utils::.OStype() ==
    "unix", 1, 3)],
  types = c("obs", "predict0", "predict1", "predict2", "iff"),
  targetdir = NULL
)
}
\arguments{
\item{parts}{vector of integer determining the parts of the files to be
created: 2 = first half, 3 = first third, ..., 10 = first 10 percent, etc.}

\item{version}{one of c("unix", "unix_home", "win32", "win32_kwb")}

\item{types}{vector of file types to be created. See the default assignment
for possible items}

\item{targetdir}{full path to the target directory in which to put the files.
By default the files go into the package's example directory related to 
the selected \code{version}.}
}
\description{
Create smaller versions of the example files
}
\examples{
\dontrun{kwbGompitz:::createExampleFilesSmall(c(2, 4, 10))}

}
