% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{createStatusMatrix}
\alias{createStatusMatrix}
\title{Default Covariate Status Matrix}
\usage{
createStatusMatrix(
  strata = NULL,
  covariateNames = NULL,
  default.other = 3,
  default.length = 0,
  covariates = "Length_num",
  Data = NULL,
  column.strata = "Material_cat",
  column.length = NULL,
  note = TRUE
)
}
\arguments{
\item{strata}{names of strata (e.g. unique values in column \emph{stratum} of
masterdata)}

\item{covariateNames}{names of covariates}

\item{default.other}{default status value to be used in any cell of the
matrix, except the cells in column \emph{LENGTH}. Default: 3}

\item{default.length}{status value to be used in column \emph{LENGTH}.
Default: 0}

\item{covariates}{names of columns in \code{Data} that shall be used as
covariates}

\item{Data}{data frame with one row per inspection}

\item{column.strata}{name of column in \code{Data} containing the stratifying
variable. Default: "Material_cat", i.e. strata are built by material}

\item{column.length}{name of column in \code{Data} containing the pipe
lengths}

\item{note}{if \code{TRUE} (default) a note about the length column is given}
}
\description{
Default covariate status matrix
}
