% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{default_condition_colours}
\alias{default_condition_colours}
\title{Get default Vector of Condition Colours}
\usage{
default_condition_colours(condition_labels)
}
\arguments{
\item{condition_labels}{vector of condition class names ordered from the best
to the worst condition}
}
\value{
vector of colour codes with the conditions as element names
}
\description{
Get default Vector of Condition Colours
}
