% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_survival.R
\name{demo_survival}
\alias{demo_survival}
\title{Interactive Plot of Survival Curves}
\usage{
demo_survival(
  t = 1:100,
  alpha.0 = -1.15,
  bz0.0 = -0.88,
  bz1.0 = -2.77,
  theme = ggplot2::theme_bw()
)
}
\arguments{
\item{t}{vector of times at which to calculate the survival curves}

\item{alpha.0}{initial value of the \code{alpha} parameter}

\item{bz0.0}{initial value of the \code{bz0} parameter}

\item{bz1.0}{initial value of the \code{bz1} parameter}

\item{theme}{ggplot2-theme applied to the plots}
}
\description{
Interactive Plot of Survival Curves
}
