% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompcal.R
\name{.calibrationAvailable}
\alias{.calibrationAvailable}
\title{Check for Convergence in Calibration}
\usage{
.calibrationAvailable(calibration, strata = NULL, param = NULL)
}
\arguments{
\item{calibration}{result of Gompitz calibration as provided by
\code{\link{runGompitzCalibration}}}

\item{strata}{vector of strata for which convergence is checked}

\item{param}{if not NULL, this structure, representing the content of the
"param.txt" file, is used instead of \code{calibration} to check for
convergence}
}
\value{
named logical vector with as many elements as there are in
  \emph{strata} each of which indicates if convergence was achieved for the
  corresponding stratum
}
\description{
Check for Convergence in Calibration
}
