% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{.fileContentStrategy0}
\alias{.fileContentStrategy0}
\title{Content for Strategy File 0}
\usage{
.fileContentStrategy0(range.years, condition.labels)
}
\arguments{
\item{range.years}{vector of two integer numbers: first and last year of
simulation}

\item{condition.labels}{vector of condition labels}
}
\description{
Content for Strategy File 0
}
