% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{.fileContentStrategy1}
\alias{.fileContentStrategy1}
\title{Content for Strategy File 1}
\usage{
.fileContentStrategy1(rehabilitation.costs, annual.total.length)
}
\arguments{
\item{rehabilitation.costs}{list of rehabilitation costs per condition. The
names of the list elements are the condition labels and the values of the
list elements are the corresponding rehabilitation costs. Example: list(C4
= 200, C3 = 300, C2 = 400, C1 = 500)}

\item{annual.total.length}{list of annual total lengths to be rehabilitated.
The names of the list elements are the year numbers and the values of the
list elements are the lenghts to be rehabilitated in the corresponding
year. Example: list("2005" = 100, "2006" = 110, "2007" = 120)}
}
\description{
Content for Strategy File 1
}
