% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{.fileContentStrategy1or2}
\alias{.fileContentStrategy1or2}
\title{File Content for Strategy 1 or 2}
\usage{
.fileContentStrategy1or2(rehabilitation.costs,
  annual.total.length.or.budget)
}
\arguments{
\item{rehabilitation.costs}{list of rehabilitation costs per condition. The
names of the list elements are the condition labels and the values of the
list elements are the corresponding rehabilitation costs. Example:
list(C4=200, C3=300, C2=400, C1=500)}

\item{annual.total.length.or.budget}{list of annual total lengths to be
rehabilitated or annual total budgets.}
}
\description{
File Content for Strategy 1 or 2
}
