% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{.fileContentStrategy2}
\alias{.fileContentStrategy2}
\title{Content for Strategy File 2}
\usage{
.fileContentStrategy2(rehabilitation.costs, annual.total.budget)
}
\arguments{
\item{rehabilitation.costs}{list of rehabilitation costs per condition. The
names of the list elements are the condition labels and the values of the
list elements are the corresponding rehabilitation costs. Example:
list(C4=200, C3=300, C2=400, C1=500)}

\item{annual.total.budget}{list of annual total budget of rehabilitation
operations. The names of the list elements are the year numbers and the
values of the list elements are the annual total budgets for the
corresponding year. Example: list("2005"=100000, "2006"=110000,
"2007"=120000)}
}
\description{
Content for Strategy File 2
}
