% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{.fileContentStrategy3}
\alias{.fileContentStrategy3}
\title{Content for Strategy File 3}
\usage{
.fileContentStrategy3(
  condition.labels,
  rehabilitation.costs,
  max.tol.prop.of.length,
  range.years,
  target.year
)
}
\arguments{
\item{condition.labels}{vector of condition labels Example: c("C4", "C3",
"C2", "C1")}

\item{rehabilitation.costs}{vector of rehabilitation costs per condition in
the order of the corresponding condition labels in condition.labels 
Example: c(200, 300, 400, 500)}

\item{max.tol.prop.of.length}{vector of maximum tolerated proportion of
network length in each condition in the order of the corresponding
condition labels in condition.labels. Example: c(1.0, 0.7, 0.4, 0.05)}

\item{range.years}{two element vector containing the first and last year of
simulation}

\item{target.year}{year at which the proportions of the network length in
each condition must have been brought just below their maximum tolerated
value}
}
\description{
Content for Strategy File 3
}
