% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompcal.R
\name{.getObservationByCondition}
\alias{.getObservationByCondition}
\title{Number of observations}
\usage{
.getObservationByCondition(calibration)
}
\arguments{
\item{calibration}{calibration result as retrieved by
\code{\link{runGompitzCalibration}}}
}
\description{
Extract number of observations from calibration result
}
