% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.orderByWeightedProbabilities}
\alias{.orderByWeightedProbabilities}
\title{probability order 2}
\usage{
.orderByWeightedProbabilities(
  probabilities,
  weight = seq_len(ncol(probabilities))
)
}
\arguments{
\item{probabilities}{matrix of probabilities}

\item{weight}{vector of weights with as many elements as there are columns in
probabilities}
}
\description{
order by weighted probabilities
}
