% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.runModuleInDirectory}
\alias{.runModuleInDirectory}
\title{.runModuleInDirectory}
\usage{
.runModuleInDirectory(
  module,
  target.dir,
  input.file,
  sep,
  ...,
  verbose = 1,
  show.error = TRUE
)
}
\arguments{
\item{module}{module}

\item{target.dir}{target.dir}

\item{input.file}{input.file}

\item{sep}{sep}

\item{...}{additional arguments passed to kwb.utils::runInDirectory}

\item{verbose}{verbose (default: 1 )}

\item{show.error}{show.error (default: TRUE)}
}
\value{
???
}
\description{
.runModuleInDirectory
}
