% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{exampleCalibration}
\alias{exampleCalibration}
\title{Calibration Object of the Gompitz Example}
\usage{
exampleCalibration(VERSION = getOperatingSystemType())
}
\arguments{
\item{VERSION}{one of "unix", "unix_home", "win32", "win32_kwb"}
}
\description{
This function returns a calibration object (as returned by
  \code{\link{runGompitzCalibration}}), corresponding to the example provided
  with the Gompitz software package
}
