% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{exampleFile}
\alias{exampleFile}
\title{Get Path to Example File}
\usage{
exampleFile(filename = "none", ..., dbg = TRUE)
}
\arguments{
\item{filename}{name of the example file}

\item{\dots}{passed to \code{\link{getDefaultPaths}}}

\item{dbg}{if \code{TRUE}, debug messages are shown, else not.}
}
\value{
full path to the example file
}
\description{
Get the full path to one of the example files provided with the GompitZ
  software package of which copies are available in this R package.
}
