% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{get_label}
\alias{get_label}
\title{Provide Label for Plot in selected Language}
\usage{
get_label(element = NULL, lng = "en")
}
\arguments{
\item{element}{one of "title", "ylab", "ylab.rel", "ylab.dev", "title.legend"}

\item{lng}{language acronym, "de" for German or "en" for English}
}
\value{
???
}
\description{
Provide Label for Plot in selected Language
}
