% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleCreate.R
\name{get_or_create_target_dir}
\alias{get_or_create_target_dir}
\title{Create target directory if required and return the path to it}
\usage{
get_or_create_target_dir(version)
}
\arguments{
\item{version}{one of \code{c("unix", "unix_home", "win32", "win32_kwb")}}
}
\description{
Create target directory if required and return the path to it
}
