% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survivals.R
\name{get_survivals}
\alias{get_survivals}
\title{Get Survival Curves from a Calibration Object}
\usage{
get_survivals(calibration, stratum, t = 1:100, matrix = TRUE, ...)
}
\arguments{
\item{calibration}{calibration object as returned by
\code{\link{runGompitzCalibration}}}

\item{stratum}{name of stratum for which to return the survival curves}

\item{t}{vector of times for which to calculate the survival probabilities}

\item{matrix}{if \code{TRUE} (default) the result is returned as a matrix,
otherwise as a list.}

\item{\dots}{further arguments passed to \code{\link{survivals}}}
}
\description{
Get Survival Curves from a Calibration Object
}
\seealso{
\code{\link{survivals}}
}
