% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompcal.R
\name{getCalibration}
\alias{getCalibration}
\title{Get a full Calibration Object from File}
\usage{
getCalibration(file_calib, file_param = NULL, digits = 5)
}
\arguments{
\item{file_calib}{full path to calibration file \code{calibr.txt}}

\item{file_param}{full path to file \code{param.txt}}

\item{digits}{passed to \code{kwbGompitz:::compareEstimates}}
}
\description{
Get a full calibration object as returned by
\code{\link{runGompitzCalibration}}, but not by calling this function but by
just reading the file \code{calibr.txt}
}
