% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompcal.R
\name{getCalibrationFile}
\alias{getCalibrationFile}
\title{Source File of Calibration Result}
\usage{
getCalibrationFile(calib.result)
}
\arguments{
\item{calib.result}{calibration result of Gompitz calibration as provided by
\code{\link{runGompitzCalibration}}}
}
\description{
Source File of Calibration Result
}
