% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{getCalibrationParameters}
\alias{getCalibrationParameters}
\title{Get parameters from Calibration Object}
\usage{
getCalibrationParameters(
  calibration = exampleCalibration(),
  remove_non_calibrated = TRUE
)
}
\arguments{
\item{calibration}{calibration object as returned by
\code{\link{runGompitzCalibration}}}

\item{remove_non_calibrated}{if \code{TRUE} (default) parameters for
non-calibrated strata are removed}
}
\description{
Get the calibration parameters as they are required to generate the
  param.txt file for the gompred program, from the calibration object
}
