% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{getConstants}
\alias{getConstants}
\title{Gompitz Covariate Status Constants}
\usage{
getConstants()
}
\description{
Gompitz constants regarding the status of covariates (not used, influencing
  the initial state, influencint the deterioration speed, influencing both)
}
\details{
@export
}
