% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{getDefaultPaths}
\alias{getDefaultPaths}
\title{Default Paths to GompitZ Files}
\usage{
getDefaultPaths(VERSION = getOperatingSystemType(), ...)
}
\arguments{
\item{VERSION}{one of "unix", "win32", "win32_kwb"}

\item{\dots}{arguments passed to resolve}
}
\description{
Default paths to GompitZ executable and example files
}
