% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{getExampleStatusMatrix}
\alias{getExampleStatusMatrix}
\title{Example Covariate Status Matrix}
\usage{
getExampleStatusMatrix()
}
\description{
Example Covariate Status Matrix
}
\seealso{
\code{\link{createStatusMatrix}}
}
