% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFile.R
\name{getFileContentForInputFile}
\alias{getFileContentForInputFile}
\title{Get file Content for Input File}
\usage{
getFileContentForInputFile(masterdata, covariates, covariates.status,
  condition.labels, weight, sep)
}
\arguments{
\item{masterdata}{masterdata as returned by \code{\link{composeMasterData}}}

\item{covariates}{data.frame with as many rows as there are in masterdata and
as many columns as there are covariates, containing the covariate values}

\item{covariates.status}{matrix of covariate status as retrieved by 
\code{\link{createStatusMatrix}}}

\item{condition.labels}{condition labels}

\item{weight}{Weight of the pipeline for the calibration}

\item{sep}{column separator}
}
\description{
Get file Content for Input File
}
