% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{getSampleObservations}
\alias{getSampleObservations}
\title{Random Observation Data}
\usage{
getSampleObservations(n = 100)
}
\arguments{
\item{n}{number of rows to be generated}
}
\value{
data frame with columns MATERIAL, Baujahr, DN, Pipe, Note,
  Inspection, PipeLength, before1970
}
\description{
Random Observation Data
}
