% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompred.R
\name{getStataLabels}
\alias{getStataLabels}
\title{Read Strata Labels from Observation File}
\usage{
getStataLabels(input.file, sep)
}
\arguments{
\item{input.file}{path to input fule}

\item{sep}{column separator}
}
\description{
Read Strata Labels from Observation File
}
