% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{gg_stacked_bars}
\alias{gg_stacked_bars}
\title{Generate gg-plot of stacked Bars}
\usage{
gg_stacked_bars(
  data,
  relative = TRUE,
  labels = FALSE,
  legend = c("bottom", "left", "top", "right", "none")[5],
  reverse = TRUE,
  lng = "en",
  colours = NULL
)
}
\arguments{
\item{data}{data frame with columns \code{length.abs} (absolute pipe length
in m), \code{length.rel} (relative pipe length in percent),
\code{condition}}

\item{relative}{logical indicating whether to show absolute or relative
lenghts}

\item{labels}{logical indicating whether to show labels (axes, title) are
shown.}

\item{legend}{legend position: one of \code{"bottom", "left", "top", "right",
"none"}}

\item{reverse}{logical indicating whether to stack the bars in reversed order}

\item{lng}{language acronym, "de" for German or "en" for English, that is
passed to \code{kwbGompitz:::get_label}}

\item{colours}{named vector that maps the values in \code{data$condition}
to colour names.}
}
\description{
Generate gg-plot of stacked Bars
}
