% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFile.R
\name{inputFileBody}
\alias{inputFileBody}
\title{Calibration Input File Body}
\usage{
inputFileBody(masterdata, covariates, weight, sep)
}
\arguments{
\item{masterdata}{masterdata as returned by \code{\link{composeMasterData}}}

\item{covariates}{data.frame with as many rows as there are in masterdata and
as many columns as there are covariates, containing the covariate values}

\item{weight}{Weight of the pipeline for the calibration}

\item{sep}{column separator}
}
\description{
Calibration Input File Body
}
\seealso{
\code{kwbGompitz:::inputFileHeader}
}
