% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFile.R
\name{inputFileHeader}
\alias{inputFileHeader}
\title{Calibration Input File Header}
\usage{
inputFileHeader(condition.labels, covariates, covariates.status,
  sep = ";")
}
\arguments{
\item{condition.labels}{condition labels}

\item{covariates}{data.frame containing covariates. Needed to determine if
covariates are numeric or categorical}

\item{covariates.status}{matrix of covariate status as retrieved by 
\code{\link{createStatusMatrix}}}

\item{sep}{column separator}
}
\description{
Calibration Input File Header
}
\seealso{
\code{kwbGompitz:::inputFileBody}
}
