% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{marginal_survival}
\alias{marginal_survival}
\title{Condition state marginal survival function surv(j,t,theta,Z)}
\usage{
marginal_survival(t, alpha, bz0, bz1, s, version = 1)
}
\arguments{
\item{t}{time}

\item{alpha}{parameter "alpha"}

\item{bz0}{parameter "bz0"}

\item{bz1}{parameter "bz1"}

\item{s}{standard deviation?}

\item{version}{1, 2, or 3}
}
\description{
alpha = theta[cond]
Integr P{Y(t) <= j | theta, Z, IFF} . phi(IFF) . dIFF
Gauss-Legendre Integration on [a,b]
}
