% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{marginal_survival_v2}
\alias{marginal_survival_v2}
\title{Marginal Survival (v2)}
\usage{
marginal_survival_v2(t, alpha, bz0, bz1, s)
}
\arguments{
\item{t}{t}

\item{alpha}{alpha}

\item{bz0}{bz0}

\item{bz1}{bz1}

\item{s}{s}
}
\value{
????
}
\description{
Marginal Survival (v2)
}
\keyword{internal}
