% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{matrixToLongDataFrame}
\alias{matrixToLongDataFrame}
\title{Convert Data in wide View to long View}
\usage{
matrixToLongDataFrame(x, columnName = deparse(substitute(x)))
}
\arguments{
\item{x}{data frame or list of data frames}

\item{columnName}{column name in result data frame}
}
\description{
Convert Data in wide View to long View
}
