% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_survival.R
\name{multicall}
\alias{multicall}
\title{Call a Function with all Combinations of Argument Ranges}
\usage{
multicall(FUN, ..., fix. = names(formals(FUN))[1],
  max.combinations = 1000)
}
\arguments{
\item{FUN}{function to be called}

\item{...}{scalars or vectors named as the arguments that are accepted by 
\code{FUN}. From all objects with a name that is not in \code{fix.} 
combinations of their values are created and \code{FUN} is called with each
of these combinations.}

\item{fix.}{names of arguments to be kept constant}

\item{max.combinations}{maximum number of argument combinations to be created
at maximum}
}
\description{
Call a function with all possible combinations of argument ranges
}
\examples{
bmi <- function(mass, height) round(mass / (height * height), 1)

kwbGompitz:::multicall(
  bmi, mass = 60:70, height = seq(1.7, 1.8, 0.1), fix. = NULL
)

}
