% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_survival.R
\name{multiplot_survival}
\alias{multiplot_survival}
\title{Plot several Survival Curves}
\usage{
multiplot_survival(
  t,
  alpha = 0,
  bz0 = 0,
  bz1 = 0,
  span = rep(TRUE, 3),
  theme = ggplot2::theme_bw()
)
}
\arguments{
\item{t}{vector of times for which to calculate the survival curves}

\item{alpha}{parameter alpha of the survival function}

\item{bz0}{parameter bz0 of the survival function}

\item{bz1}{parameter bz1 of the survival function}

\item{span}{vector of logical of length three indicating for each of the 
parameters \code{alpha}, \code{bz0}, \code{bz1} if they are to be varied or
to be kept fix}

\item{theme}{ggplot2-theme applied to the plots}
}
\description{
Plot several Survival Curves
}
