% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{open_file_if_not_on_unix}
\alias{open_file_if_not_on_unix}
\title{Open File if not on UNIX}
\usage{
open_file_if_not_on_unix(file)
}
\arguments{
\item{file}{file}
}
\value{
open file
}
\description{
Open File if not on UNIX
}
\keyword{internal}
