% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{parameterLines}
\alias{parameterLines}
\title{Generate text lines for param.txt}
\usage{
parameterLines(parameters, sep = ";")
}
\arguments{
\item{parameters}{list structure containing calibration parameters as
provided by \code{kwbGompitz:::readParameters}}

\item{sep}{column separator. Default: ";"}
}
\value{
vector of character representing the lines of param.txt, containing
  the calibration parameters as generated by /code{gompcal}
}
\description{
Convert the list structure containing calibration parameters as retrieved by
\code{kwbGompitz:::readParameters} to a vector of character representing the
text lines to be written to file
}
