% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_survival.R
\name{parameterplot}
\alias{parameterplot}
\title{Plot multiple Survival Curves}
\usage{
parameterplot(t, alpha = 0, bz0 = 0, bz1 = 0, xlim = c(0, 1.2 *
  max(t)))
}
\arguments{
\item{t}{vector of times for which to calculate the survival curves}

\item{alpha}{vector of parameters alpha to be passed to the survival function}

\item{bz0}{vector of parameters bz0 to be passed to the survival function}

\item{bz1}{vector of parameters bz1 to be passed to the survival function}

\item{xlim}{limits of the x axis}
}
\description{
Plot multiple Survival Curves
}
