% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG_2.R
\name{plot_pipe_conditions}
\alias{plot_pipe_conditions}
\title{Plot the Pipe Conditions over Time}
\usage{
plot_pipe_conditions(
  x,
  column_instyear = "instyear",
  column_ident = "ident",
  column_inspyear = "inspyear",
  column_condition = "condition",
  colour_values = c(C4 = "green", C3 = "yellow", C2 = "orange", C1 = "red", "grey"),
  y_labels = FALSE,
  facet_by = sprintf(kwb.utils::underscoreToPercent("~(10*as.integer(_s/10))"),
    column_instyear),
  by_pipe = FALSE
)
}
\arguments{
\item{x}{data frame}

\item{column_instyear}{name of column containing the installation year}

\item{column_ident}{name of column containing the pipe identifier}

\item{column_inspyear}{name of column containing the inspection year}

\item{column_condition}{name of column containing the pipe condition}

\item{colour_values}{named vector of colour names. The names must refer to
the values in \code{column_condition}.}

\item{y_labels}{logical. If \code{TRUE} tick marks and labels are plotted
on the y axis, else not.}

\item{facet_by}{character string representing a formula to be used to create
facets with \code{\link[ggplot2]{facet_wrap}}. Set to \code{NULL} if no
facets are desired.}

\item{by_pipe}{if \code{TRUE} one plot is created for each pipe and a list
of these plots is returned}
}
\description{
Plot the Pipe Conditions over Time
}
\examples{
\dontrun{
x <- kwbGompitz:::readObservations(kwbGompitz::exampleFile("obs.txt"))
plot_pipe_conditions(x[1:100, ])
}

}
