% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG_2.R
\name{plot_prediction_by_pipe}
\alias{plot_prediction_by_pipe}
\title{Plot Condition Probabilities by Pipe}
\usage{
plot_prediction_by_pipe(
  prediction,
  prefix = "prob",
  pipe_ids = unique(kwb.utils::selectColumns(prediction, "pipe_id")),
  width = 1
)
}
\arguments{
\item{prediction}{data frame as returned by \code{\link{runGompitzPrediction}}}

\item{prefix}{prefix of column names containing the probabilities. Default: "prob"}

\item{pipe_ids}{vector of pipe IDs for which a plot is to be generated. By 
default all available pipes are considered!}

\item{width}{passed to \code{\link[ggplot2]{geom_col}} (default: 1)}
}
\value{
???
}
\description{
List of ggplots with each plot representing the evolution of 
condition probabilities over time for one pipe
}
