% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{plot_stacked_bars}
\alias{plot_stacked_bars}
\title{Generate gg-plot of Stacked Bars}
\usage{
plot_stacked_bars(x, x.sd = NULL, reverse = TRUE, ...)
}
\arguments{
\item{x}{matrix containing the bar heights}

\item{x.sd}{optional. Standard deviations.}

\item{reverse}{logical indicating whether to reverse the stack order}

\item{\dots}{additional arguments passed to \code{\link{gg_stacked_bars}}}
}
\description{
Generate gg-plot of Stacked Bars
}
\examples{

values <- c(
  0, 2, 0, 0, 0,
  1, 0, 0, 2, 0,
  0, 0, 0, 2, 5,
  0, 0, 3, 0, 0
)

conditionStat <- matrix(values, nrow = 4, byrow = TRUE, dimnames = list(
  condition = paste0("prob", 1:4),
  year = 2001:2005
))

# Get the base plot
baseplot <- kwbGompitz::plot_stacked_bars(conditionStat, legend = "right")

# Show the base plot
baseplot

# Modify the base plot (titles, axis titles, legend title, legend content)
baseplot + ggplot2::labs(
  x = "Jahr", y = "Anteil Kanaele in \%", title = "Zustandsverteilung",
  subtitle = "Netzwerk: Berlin"
) +
ggplot2::guides(fill = ggplot2::guide_legend("Zustandsklasse")) +
ggplot2::scale_fill_manual(
  values = c("darkgreen", "yellow", "darkorange3", "red3"),
  labels = c("good", "ok", "not so good", "bad")
)

}
