% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_survival_curves}
\alias{plot_survival_curves}
\title{Plot Survival Curves}
\usage{
plot_survival_curves(calibration = exampleCalibration(),
  stratum = NULL, marginal = TRUE, t = 1:100, col = NULL, ...,
  args = NULL)
}
\arguments{
\item{calibration}{calibration object}

\item{stratum}{name of stratum}

\item{marginal}{logical. If \code{TRUE} (default) the marginal survival
curves are plotted, otherwise the non-marginal survival curves}

\item{t}{vector of times at which to calculate the survival curves}

\item{col}{vector of colour names, named according to the conditions as
stored in \code{attr(calibration, "parameters")$conditions}}

\item{\dots}{arguments passed to \code{\link[kwb.plot]{plot_curve_areas_gg}},
such as \code{legend}, \code{line_colour}}

\item{args}{arguments passed to get_survivals}
}
\description{
Plot Survival Curves
}
\examples{
\dontrun{
# Get the example calibration provided with the Gompitz software
calibration <- kwbGompitz::exampleCalibration()

# Generate one ggplot2-object for each calibrated stratum
(plots_1 <- plot_survival_curves(calibration))

# By default the marginal survival curves are shown. You may set marginal to
# FALSE to get the non-marginal survival curves
(plots_2 <- plot_survival_curves(calibration, marginal = FALSE))

# Compare the plots
gridExtra::grid.arrange(plots_1[[1]], plots_2[[1]], plots_1[[2]], plots_2[[2]])
}

}
