% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCalibration1}
\alias{plotCalibration1}
\title{Plot Calibration Result}
\usage{
plotCalibration1(calib, to.pdf = FALSE)
}
\arguments{
\item{calib}{calibration result as retrieved by
\code{\link{runGompitzCalibration}}}

\item{to.pdf}{if \code{TRUE} the plot is directed into a temporary PDF file}
}
\description{
Plot Calibration Result
}
