% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCalibration2}
\alias{plotCalibration2}
\title{Plot Calibration Result (2)}
\usage{
plotCalibration2(calib, to.pdf = FALSE)
}
\arguments{
\item{calib}{calibration result as retrieved by
\code{\link{runGompitzCalibration}}}

\item{to.pdf}{if \code{TRUE} the plot is directed into a temporary PDF file}
}
\description{
Plot Calibration Result (2)
}
