% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{plotPrediction}
\alias{plotPrediction}
\title{Plot the Result of a Prediction}
\usage{
plotPrediction(
  prediction,
  legend_pos = c("bottom", "left", "top", "right", "none")[5],
  do_print = TRUE
)
}
\arguments{
\item{prediction}{data frame with columns \code{prob1}, \code{prob2}, ...,
\code{pipeLength}, \code{InspectionYear}}

\item{legend_pos}{legend position: one of \code{"bottom", "left", "top",
"right", "none"}}

\item{do_print}{logical indicating whether to actually print the plot to
the current graphical device.}
}
\description{
Plot the distribution of condition classes by predicted year
}
\examples{

prediction <- data.frame(
  prob1 = c(0, 1, 0, 0.0, 0),
  prob2 = c(1, 0, 0, 0.5, 0),
  prob3 = c(0, 0, 0, 0.5, 1),
  prob4 = c(0, 0, 1, 0.0, 0),
  pipeLength = 1:5,
  InspectionYear = seq(2001, 2005)
)

kwbGompitz:::plotPrediction(prediction, legend_pos = "right")
}
