% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPredictionByYear}
\alias{plotPredictionByYear}
\title{Plot Prediction by Year to temp. PDF File}
\usage{
plotPredictionByYear(prediction, to.pdf = TRUE,
  FUN.name = "orderByWeightedProbabilities", column.year = "year")
}
\arguments{
\item{prediction}{prediction result as provided by e.g.
\code{\link{runGompitzPrediction}}}

\item{to.pdf}{if \code{TRUE} (default) the plot goes into a PDF file}

\item{FUN.name}{name of function to be used to order the matrix of
probabilities. Available functions: "orderByWeightedProbabilities",
"orderByMostProbClassAndDecreasingProb"; default:
"orderByWeightedProbabilities"}

\item{column.year}{name of column containing the year}
}
\description{
Plot prediction by year to temporary PDF file
}
\seealso{
\code{\link{runGompitzPrediction}}
}
