% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPredictionForYear}
\alias{plotPredictionForYear}
\title{Plot Prediction for the given Year}
\usage{
plotPredictionForYear(
  res,
  year,
  FUN.name = "orderByWeightedProbabilities",
  main = NULL,
  column.year = "year"
)
}
\arguments{
\item{res}{prediction result as provided by e.g.
\code{\link{runGompitzPrediction}}}

\item{year}{year number for which results are to be filtered an plot is to be
generated}

\item{FUN.name}{name of function to be used to order the matrix of
probabilities. Available functions: "orderByWeightedProbabilities",
"orderByMostProbClassAndDecreasingProb"; default:
"orderByWeightedProbabilities"}

\item{main}{main title of plot}

\item{column.year}{name of column containing the year}
}
\description{
Plot Prediction for the given Year
}
\seealso{
\code{\link{runGompitzPrediction}}
}
