% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{prepare_for_gg_stacked_bars}
\alias{prepare_for_gg_stacked_bars}
\title{Prepare data for plot_stacked_bars()}
\usage{
prepare_for_gg_stacked_bars(x, x.sd = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{x.sd}{standard deviation}
}
\description{
Prepare data for plot_stacked_bars()
}
