% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gompitz.calibration}
\alias{print.gompitz.calibration}
\title{Print a GompitZ Calibration Structure}
\usage{
\method{print}{gompitz.calibration}(x, ...)
}
\arguments{
\item{x}{list of class "gompitz.calibration" as returned by
\code{\link{runGompitzCalibration}}}

\item{\dots}{further arguments (not used)}
}
\description{
Print a GompitZ Calibration Structure
}
