% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gompitz_stratum_calib}
\alias{print.gompitz_stratum_calib}
\title{Print the Calibration for one Stratum}
\usage{
\method{print}{gompitz_stratum_calib}(x, ...)
}
\arguments{
\item{x}{list of class "gompitz_stratum_calib" as contained in a calibration
object returned by \code{\link{runGompitzCalibration}}}

\item{\dots}{further arguments (not used)}
}
\description{
Print the Calibration for one Stratum
}
