% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printConvergence}
\alias{printConvergence}
\title{Print the Convergence Information}
\usage{
printConvergence(x)
}
\arguments{
\item{x}{\code{NULL} or a list with elements \code{num.iterations},
\code{log.likelihood}, \code{covariances}, \code{estimates},}
}
\value{
Print the convergence information
}
\description{
Print the Convergence Information
}
