% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCalibration.R
\name{readCalibration}
\alias{readCalibration}
\title{Read Calibration Result}
\usage{
readCalibration(file, verbose = 1)
}
\arguments{
\item{file}{path to gompcal result file "calibr.txt"}

\item{verbose}{integer number specifying the level of verbosity}
}
\description{
Read calibration result from given file
}
