% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFile.R
\name{readObservations}
\alias{readObservations}
\title{Read Observation Data from File in GompitZ-Format}
\usage{
readObservations(file = exampleFile("obs.txt"), sep = ";",
  dbg = TRUE)
}
\arguments{
\item{file}{full path to text file in the format that is required by GompitZ
(see "The GompitZ Tool User's Guide")}

\item{sep}{column separator}

\item{dbg}{if \code{TRUE} debug messages are shown, else not.}
}
\description{
Read Observation Data from File in GompitZ-Format
}
