% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{readParameters}
\alias{readParameters}
\title{Read Parameters from "param.txt"}
\usage{
readParameters(file, sep = ";", dbg = FALSE)
}
\arguments{
\item{file}{full path to parameter file "param.txt", generated by gompcal.exe}

\item{sep}{column separator, default: ";"}

\item{dbg}{if \code{TRUE} (default) debug messages are shown}
}
\value{
list with sections \code{conditions}, /code{strata},
  /code{covariates}, /code{categoryLevels}, /code{byStratum},
}
\description{
Read parameters from param.txt into list structure
}
