% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrediction.R
\name{readPredictionFile}
\alias{readPredictionFile}
\title{Read Prediction Result File}
\usage{
readPredictionFile(file, sep = ";",
  stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{file}{full path to prediction result file (predict<n>.txt) with <n>
being the strategy number}

\item{sep}{column separator, default: ";"}

\item{stringsAsFactors}{if \code{TRUE} character columns will appear as
factors in the result, passed to \code{\link{data.frame}}}
}
\description{
Read Prediction Result File
}
