% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrediction.R
\name{readPredictionFileWithDataTable}
\alias{readPredictionFileWithDataTable}
\title{Read Predictions File with data.table}
\usage{
readPredictionFileWithDataTable(file, sep = ";")
}
\arguments{
\item{file}{file}

\item{sep}{sep (default: ";")}
}
\value{
???
}
\description{
Read Predictions File with data.table
}
\keyword{internal}
