% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runGompcalInDirectory}
\alias{runGompcalInDirectory}
\title{Run gompcal.exe in given Directory}
\usage{
runGompcalInDirectory(target.dir = tempGompitzDir(),
  input.file = exampleFile("obs.txt"), sep = ";", ...)
}
\arguments{
\item{target.dir}{path to target directory}

\item{input.file}{path to input file}

\item{sep}{column separator}

\item{\dots}{arguments that are passed to .runModuleInDirectory, such as
\code{verbose} or \code{show.error}}
}
\description{
Run gompcal.exe in given Directory
}
