% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompcal.R
\name{runGompitzCalibration}
\alias{runGompitzCalibration}
\title{Perform GompitZ Calibration}
\usage{
runGompitzCalibration(input.data, subset = NULL, verbose = 1,
  sep = ";", digits = 3, VERSION = getOperatingSystemType())
}
\arguments{
\item{input.data}{prepared input data as retrieved by
\code{\link{composeGompitzInputData}}}

\item{subset}{indexes of rows in \code{input.data} to be used for 
calibration. If \code{NULL} (default), all rows are used.}

\item{verbose}{verbosity level}

\item{sep}{column separator}

\item{digits}{round estimates read from calibr.txt and param.txt,
respectively, to this number of (significant) digits before comparing them}

\item{VERSION}{name of subdirectory in package containing the binary files to
be executed. Possible values: "unix", "win32", "win32_kwb"}
}
\description{
Perform GompitZ Calibration
}
\examples{
# For an example, see the Tutorial vignette "How to Use the Package"

}
