% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_survival.R
\name{spanIf}
\alias{spanIf}
\title{Span a Vector around x if Condition is met}
\usage{
spanIf(condition, x, digits = 4)
}
\arguments{
\item{condition}{logical. If \code{TRUE} a vector around \code{x} is spanned}

\item{x}{value around which to span a vector (or not if \code{condition} is
\code{FALSE})}

\item{digits}{number of digits to which the values are rounded}
}
\description{
Span a Vector around x if Condition is met
}
