% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\docType{data}
\name{SQRT_2_PI}
\alias{SQRT_2_PI}
\title{Mathematical Constant \code{sqrt(2 * pi)}}
\format{An object of class \code{numeric} of length 1.}
\usage{
SQRT_2_PI
}
\description{
Mathematical Constant \code{sqrt(2 * pi)}
}
\keyword{datasets}
