% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{standard_survival}
\alias{standard_survival}
\title{Weibull condition state survival function}
\usage{
standard_survival(
  alpha,
  t,
  bz1,
  bz0 = 0,
  offset = t * exp(bz1) + bz0,
  limits = NULL
)
}
\arguments{
\item{alpha}{parameter "alpha"}

\item{t}{time}

\item{bz1}{parameter "bz1"}

\item{bz0}{parameter "bz0"}

\item{offset}{in order not to recalculate the following expression each time
again, its result can be given here: \code{= bz0 + t * exp(bz1)}. If given,
the arguments \code{t, bz0, and bz1} can be omitted, otherwise they are
required and the offset is calculated according to the above expression.}

\item{limits}{numeric vector of two elements giving the minimum and maximum
value to which the result shall be restricted. If not given or \code{NULL}
the result will not be restricted to a value range.}
}
\description{
Weibull condition state survival function
}
