% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{strategyFileContent}
\alias{strategyFileContent}
\title{Content for Strategy File}
\usage{
strategyFileContent(strategy, ...)
}
\arguments{
\item{strategy}{strategy identifier. Must be one of 0 (do nothing), 1
(length-driven strategy), 2 (budget-driven strategy), or 3 (optimised
strategy).")}

\item{\dots}{arguments passed to the corresponding
\code{.fileContentStrategy} functions, see hidden functions 
\code{kwbGompitz:::.fileContentStrategy0}, 
\code{kwbGompitz:::.fileContentStrategy1}, 
\code{kwbGompitz:::.fileContentStrategy2}, 
\code{kwbGompitz:::.fileContentStrategy3}}
}
\value{
character vector of length one representing the file content of the 
  strategy file
}
\description{
Create the text content for a gompred strategy file
}
