% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{summary_generic_sim}
\alias{summary_generic_sim}
\title{Aggregate simulated Condition Classes}
\usage{
summary_generic_sim(x, column.group.by, column.length)
}
\arguments{
\item{x}{data frame}

\item{column.group.by}{name of column in \code{x} by which values to group by}

\item{column.length}{name of column in \code{x} containing the pipe lengths}
}
\description{
Aggregate simulated Condition Classes
}
