% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survivals.R
\name{survivals}
\alias{survivals}
\title{Get Values of (Standard or Marginal) Survival Curves}
\usage{
survivals(
  t = 0:99,
  alpha,
  bz1,
  bz0,
  s = NULL,
  marginal = !is.null(s),
  matrix = TRUE,
  set_attributes = FALSE,
  interpol_info = NULL,
  ...
)
}
\arguments{
\item{t}{numeric vector of times (ages)}

\item{alpha}{numeric vector of alpha-parameter(s)}

\item{bz1}{bz1 parameter}

\item{bz0}{bz0 parameter}

\item{s}{passed to \code{\link{marginal_survival}}}

\item{marginal}{if \code{TRUE} the marginal survival curve with \code{s} as
standard deviation is calculated instead of the standard survival curve.
By default \code{marginal} is \code{TRUE} if \code{s} is not \code{NULL}.}

\item{matrix}{if \code{TRUE} and the length of \code{alpha} is greater than
one the result is a matrix with each row representing one \code{alpha}
value and each column representing a time. Otherwise the result is a list
with each list element representing one \code{alpha} value}

\item{set_attributes}{if \code{TRUE} (the default is \code{FALSE}) an
attribute \code{args} containing the arguments given to this function is
set in the result}

\item{interpol_info}{if not \code{NULL} it is expected to be a list
containing the elements \code{start} (start year of interpolation) and
\code{length} (duration of interpolation in years)}

\item{\dots}{further arguments passed to \code{\link{marginal_survival}} or
\code{\link{standard_survival}}}
}
\description{
Get Values of (Standard or Marginal) Survival Curves
}
