% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{tempGompitzDir}
\alias{tempGompitzDir}
\title{Create temporary gompitz directory}
\usage{
tempGompitzDir(verbose = 1)
}
\arguments{
\item{verbose}{integer value determining the level of verbosity}
}
\description{
Create temporary gompitz directory
}
