% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{test_marginal_survival}
\alias{test_marginal_survival}
\title{Test Marginal Survival}
\usage{
test_marginal_survival(t = 0:99)
}
\arguments{
\item{t}{t (default: 0:99)}
}
\value{
???
}
\description{
Test Marginal Survival
}
