% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{to_theme_legend}
\alias{to_theme_legend}
\title{Generate ggplot Theme for Legend}
\usage{
to_theme_legend(legend)
}
\arguments{
\item{legend}{logical indicating whether to put a legend or not or a string
giving the legend position ("left", "right", "top", "bottom")}
}
\description{
Generate ggplot Theme for Legend
}
