% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGG.R
\name{toColours}
\alias{toColours}
\title{Condition Class(es) to Colour String(s)}
\usage{
toColours(x)
}
\arguments{
\item{x}{vector of condition class numbers or labels "prob1", "prob2", ...}
}
\description{
Condition Class(es) to Colour String(s)
}
