% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFile.R
\name{toStatusMatrix}
\alias{toStatusMatrix}
\title{Convert Text Lines to Status Matrix}
\usage{
toStatusMatrix(textlines, sep = ";", order_rows = FALSE)
}
\arguments{
\item{textlines}{vector of character linees as they appear in the GompitZ
input file header (only the lines defining the status matrix)}

\item{sep}{column separator. Default: ";"}

\item{order_rows}{if \code{TRUE} (default) the rows are ordered by their
name}
}
\description{
Convert text lines to status matrix, as e.g. required by 
\code{kwb.rsproto::configure}
}
\examples{
\dontrun{
file <- kwbGompitz::exampleFile("obs.txt")
kwbGompitz:::toStatusMatrix(textlines = readLines(file, 8)[-(1:2)])
}
}
