% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{underline}
\alias{underline}
\title{Create a String to be used as an Underline}
\usage{
underline(n = 10, char = "-")
}
\arguments{
\item{n}{number of characters}

\item{char}{character used for the underline}
}
\description{
Create a String to be used as an Underline
}
