% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompcal.R
\name{writeInputFile}
\alias{writeInputFile}
\title{Write the Input File}
\usage{
writeInputFile(textlines, file, verbose = 1)
}
\arguments{
\item{textlines}{vector of text lines to be written to \code{file}}

\item{file}{path to the file to be written}

\item{verbose}{integer number specifying the verbosity level. If this is a 
positive value, debug messages are shown.}
}
\description{
Write the given character vector into the input file at given path
}
