% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runGompredInDirectory}
\alias{runGompredInDirectory}
\title{Run gompred.exe in given Directory}
\usage{
runGompredInDirectory(target.dir = tempdir(),
  input.file = exampleFile("obs.txt"), sep = ";", strategy = 0, ...)
}
\arguments{
\item{target.dir}{path to target directory}

\item{input.file}{path to input file}

\item{sep}{column separator}

\item{strategy}{integer number specifying the strategy to be applied}

\item{\dots}{arguments that are passed to .runModuleInDirectory, such as
\code{verbose} or \code{show.error}}
}
\description{
Run gompred.exe in given Directory
}
